// drizzle/schema.d.ts
export declare const orders: import("drizzle-orm/mysql-core").MySqlTableWithColumns<{
    name: "orders";
    schema: undefined;
    columns: {
// ...
        status: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "status";
            tableName: "orders";
            dataType: "string";
            columnType: "MySqlEnumColumn";
            data: "pending" | "paid" | "processing" | "shipped" | "delivered" | "cancelled";
            driverParam: string;
            notNull: true;
            hasDefault: true;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: ["pending", "paid", "processing", "shipped", "delivered", "cancelled"];
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
// ...
    };
    dialect: "mysql";
}>;
// ...
export declare const checkoutSessions: import("drizzle-orm/mysql-core").MySqlTableWithColumns<{
    name: "checkout_sessions";
    schema: undefined;
    columns: {
// ...
        status: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "status";
            tableName: "checkout_sessions";
            dataType: "string";
            columnType: "MySqlEnumColumn";
            data: "failed" | "pending" | "paid";
            driverParam: string;
            notNull: true;
            hasDefault: true;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: ["pending", "paid", "failed"];
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
// ...
    };
    dialect: "mysql";
}>;
// ...
export type CheckoutSession = typeof checkoutSessions.$inferSelect;